'提交修改'

+++ projects/malls-admin/core/src/lib/builder/d-workspace.component.ts
@@ -1,11 +1,13 @@
-import {Component, TemplateRef, ViewChild} from '@angular/core';
+import {Component, OnDestroy, OnInit, TemplateRef} from '@angular/core';
 import {DesignService} from '../public/service/design.service';
-import {DMainComponent} from './d-main.component';
-import {Component as UIComponent, Setting, View} from '../public/model';
+import {Cell, Component as UIComponent, Row, View} from '../public/model';
 import {ModalService} from '../public/service/modal.service';
-import {SettingService} from '../public/service/setting.service';
 import {ViewService} from '../public/service/view.service';
 import {RouterService} from '../public/service/router.service';
+import {NavigationEnd, Router, RouterEvent} from '@angular/router';
+import {Subscription} from 'rxjs';
+import {I18nService} from '../public/service/i18n.service';
+import {ComponentManager} from '../registry/component-manager';
 
 
 export interface ComponentGroup {
@@ -14,83 +16,86 @@ export interface ComponentGroup {
     components: { type: string; name: string }[];
 }
 
-const componentGroups:ComponentGroup[] = [
-    {
-        icon:'layout',
-        name:'布局',
-        components:[
-            {type:'row',name:'栅格'},
-            // {type:'layout',name:'布局'},
-        ]
-    },
-    {
-        icon:'table',
-        name:'数据展示',
-        components:[
-            {type:'table',name:'表格'},
-            {type:'list',name:'列表'},
-            {type:'card',name:'卡片'},
-            {type:'tabset',name:'标签页'}
-        ]
-    },
-    {
-        icon:'form',
-        name:'表单控件',
-        components:[
-            {type:'form',name:'表单'},
-            // {type:'fieldset',name:'字段集'},
-            // {type:'array',name:'数组字段'},
-            // {type:'map',name:'map字段'},
-            {type:'text',name:'文本'},
-            {type:'number',name:'数字'},
-            {type:'textarea',name:'多行文本'},
-            // {type:'rich-text',name:'富文本'},
-            {type:'date',name:'日期'},
-            {type:'date-range',name:'日期区间'},
-            {type:'time',name:'时间'},
-            {type:'switch',name:'开关'},
-            {type:'slider',name:'滑动条'},
-            {type:'rate',name:'评分'},
-            {type:'select',name:'下拉选择'},
-            {type:'radio',name:'单选框'},
-            {type:'checkbox',name:'多选框'},
-            {type:'upload',name:'上传'},
-            {type:'data-picker',name:'数据引用'},
-            {type:'display-text',name:'文本展示'}
-        ]
-    }
-];
 
 @Component({
+    selector: 'd-workspace',
     templateUrl: './d-workspace.component.html',
     styleUrls: ['./d-workspace.component.less']
 })
-export class DWorkspaceComponent {
+export class DWorkspaceComponent implements OnDestroy, OnInit {
 
-    routerChild: DMainComponent;
+    view: View;
 
-    @ViewChild('modalHeader')
-    modalHeader: TemplateRef<any>;
+    rawData: any;
 
-    @ViewChild('saveBody')
-    saveBody: TemplateRef<any>;
+    un: Subscription;
 
-    setting: Setting;
-
-    groups = componentGroups;
+    groups: ComponentGroup[] = [];
 
     basePath: string;
 
-    constructor(private routerService: RouterService,
+    language: string;
+
+    constructor(private router: Router,
+                private routerService: RouterService,
                 private modalService: ModalService,
                 private designService: DesignService,
-                private settingsService: SettingService,
-                private viewService: ViewService) {
-        this.settingsService.subscribe(setting => this.setting = setting);
-
+                private viewService: ViewService,
+                private componentManager: ComponentManager,
+                i18nService: I18nService) {
         this.basePath = routerService.basePath;
-
+        this.language = i18nService.getLocale();
+        this.un = this.router.events.subscribe(this.initResource);
     }
+
+    ngOnInit(): void {
+        let registrations = this.componentManager.getAllComponentRegistrations();
+
+        this.groups.push({
+            icon:'layout',
+            name:'布局',
+            components: registrations.filter(reg=>reg.group === 'layout').map(({type,name})=>({type,name}))
+        });
+
+        this.groups.push({
+            icon:'table',
+            name:'数据展示',
+            components: registrations.filter(reg=>reg.group === 'display').map(({type,name})=>({type,name}))
+        });
+
+        this.groups.push({
+            icon:'form',
+            name:'表单控件',
+            components: this.componentManager.getAllFormControlRegistrations().map(({type,name})=>({type,name}))
+        })
+    }
+
+
+
+    ngOnDestroy(): void {
+        this.un.unsubscribe();
+    }
+
+
+    initResource = (event: RouterEvent) => {
+        if (event instanceof NavigationEnd) {
+            let path = this.routerService.getContentComponentPath().substring(2); //remove`/d`prefix
+            if(path) {
+                this.viewService.getCompatibleView(decodeURIComponent(path)).then(view => {
+                    this.view = view;
+                    this.rawData = view.data;
+                    this.view.data = UIComponent.create(view.data);
+                });
+            } else{
+                this.view = {
+                    language: this.language,
+                    path: '/please-input-unique-path',
+                    name: 'view name',
+                    data: new Row({children: [new Cell(), new Cell(), new Cell()]})
+                };
+            }
+        }
+    };
 
 
     addComponent(type: string) {
@@ -101,33 +106,55 @@ export class DWorkspaceComponent {
         }
     }
 
-    saveView() {
+
+    saveView(header: TemplateRef<any>, content: TemplateRef<any>) {
         let agent = this.modalService.create({
             nzWidth: '40%',
-            nzTitle: this.modalHeader,
-            nzContent: this.saveBody,
+            nzTitle: header,
+            nzContent: content,
             nzOnOk: () => {
                 if (this.view.name && this.view.path) {
-                    this.viewService.saveOrUpdateView(this.view);
-                    agent.destroy();
+                    if(this.rawData) {
+                        this.viewService.saveOrUpdateView(this.view);
+                        agent.destroy();
+                    }
+                    else {
+                        this.viewService.exists(decodeURIComponent(this.view.path))
+                            .then(exists=> {
+                                if (exists) {
+                                    let confirmAgent = this.modalService.confirm({
+                                        nzTitle: '强制覆盖视图',
+                                        nzContent: `视图${this.view.path}已经存在，是否直接覆盖？`,
+                                        nzOnOk: ()=>{
+                                            this.viewService.saveOrUpdateView(this.view);
+                                            confirmAgent.destroy();
+                                        }
+                                    })
+                                } else {
+                                    this.viewService.saveOrUpdateView(this.view);
+                                    agent.destroy();
+                                }
+                            })
+                    }
                 }
             }
         });
     }
 
-    revert() {
-        this.routerChild.revert();
-    }
 
     deleteView() {
-        this.viewService.deleteView(this.view.path);
+        let confirmAgent = this.modalService.confirm({
+            nzTitle: '删除视图？',
+            nzContent: `确认删除视图${this.view.path}？`,
+            nzOnOk: ()=>{
+                this.viewService.deleteView(this.view.path);
+                confirmAgent.destroy();
+            }
+        })
+
     }
 
-    activate(event: DMainComponent) {
-        this.routerChild = event;
-    }
-
-    get view(): View {
-        return this.routerChild.view;
+    revert() {
+        this.view.data = UIComponent.create(this.rawData);
     }
 }
